fix: disappear of reward and PE records
Fix the issue that the reward and PE information is not record
in the metrics.csv file, the cause the that the _dump_logs from
OnPolicyAlgorithm will call self.logger.dump to dump and clear the
record of logger, thus, the reward and PE information is gone. To
fix this bug, just disable the dump_logs in the _dump_logs function.

Dump and record after the training process is done, that is, the
self.train() function from the OnPolicyAlgorithm.learn() function.